refactor(web): asset store
